requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sentence-transformers>=5.0.0
faiss-cpu==1.8.0
torch==2.2.0
transformers>=4.40.0
python-multipart==0.0.6
numpy==1.26.4
requests==2.31.0
accelerate==0.20.3
